+++ iosApp/iosApp/Navigation/AppRouter.swift
@@ -2,7 +2,7 @@ import SwiftUI
 
 enum AppRoute: Hashable {
     case productDetail(sku: Int64)
-    case arObject(filePath: String, widthMm: Int, heightMm: Int, depthMm: Int)
+    case arObject(filePath: String, widthMm: Float, heightMm: Float, depthMm: Float)
 }
 
 final class AppRouter: ObservableObject {
